Add files via upload
Added the burger menu, added images for projects.html, navbar is consistent and so is the footer throughout all the pages, animations are working good. Next up to do is the dropdown for the table of contents, and for the pages to scroll down a bit when first loaded into.